requirements.txt
opencv-python
networkx
matplotlib.pyplot